third_party/blink/web_tests/platform/generic/editing/selection/drag-drop-restore-expected.txt

 

This test verifies that content of input box could be restored in 'focus' event fired by drag-and-drop.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS successfullyParsed is true

TEST COMPLETE

PASS oldValue is ""
PASS src.value is "abc"
PASS dst.value is "abc"
